+++ demo/main.gd
@@ -1,6 +1,7 @@
 extends Node3D
 
 var xr_interface : XRInterface = null
+var scene_capture: OpenXRFbSceneCaptureExtensionWrapper = null
 
 # Called when the node enters the scene tree for the first time.
 func _ready():
@@ -8,8 +9,16 @@ func _ready():
 	if xr_interface and xr_interface.is_initialized():
 		var vp: Viewport = get_viewport()
 		vp.use_xr = true
+		
+	scene_capture = OpenXRFbSceneCaptureExtensionWrapper.get_singleton()
+	scene_capture.connect("scene_capture_completed", _on_scene_capture_completed)
 
 
-# Called every frame. 'delta' is the elapsed time since the previous frame.
-func _process(delta):
-	pass
+func _on_scene_capture_completed():
+	print("Scene Capture completed")
+
+
+func _on_left_hand_button_pressed(name):
+	if name == "menu_button" and scene_capture:
+		print("Triggering scene capture")
+		scene_capture.request_scene_capture()