+++ godotopenxrmeta/libs/ovr_openxr_mobile_sdk/OpenXR/Include/openxr/metax2_detached_controllers.h
@@ -1,0 +1,34 @@
+#ifndef METAX2_DETACHED_CONTROLLERS_H_
+#define METAX2_DETACHED_CONTROLLERS_H_ 1
+
+/**********************
+This file is @generated from the OpenXR XML API registry.
+Language    :   C99
+Copyright   :   (c) Meta Platforms, Inc. and affiliates. Confidential and proprietary.
+***********************/
+
+#include <openxr/openxr.h>
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+
+#ifndef XR_METAX2_detached_controllers
+
+#define XR_METAX2_detached_controllers 1
+#define XR_METAX2_detached_controllers_SPEC_VERSION 1
+#define XR_METAX2_DETACHED_CONTROLLERS_EXTENSION_NAME "XR_METAX2_detached_controllers"
+#endif /* XR_METAX2_detached_controllers */
+
+#ifdef  XR_METAX2_DETACHED_CONTROLLERS_TAG_ALIAS
+#define XR_META_detached_controllers_SPEC_VERSION XR_METAX2_detached_controllers_SPEC_VERSION
+#define XR_META_DETACHED_CONTROLLERS_EXTENSION_NAME XR_METAX2_DETACHED_CONTROLLERS_EXTENSION_NAME
+#endif /* XR_METAX2_DETACHED_CONTROLLERS_TAG_ALIAS */
+
+
+#ifdef __cplusplus
+}
+#endif
+
+#endif